bulk-cdk-toolkit-extract-cdc,source-mysql-v2: move CDC files around
