+++ airbyte-cdk/bulk/toolkits/extract-cdc/src/main/kotlin/io/airbyte/cdk/command/CdcSourceConfiguration.kt
@@ -2,9 +2,7 @@
  * Copyright (c) 2024 Airbyte, Inc., all rights reserved.
  */
 
-package io.airbyte.cdk.read
-
-import io.airbyte.cdk.command.SourceConfiguration
+package io.airbyte.cdk.command
 
 /** Subtype of [SourceConfiguration] for CDC sources. */
 interface CdcSourceConfiguration : SourceConfiguration {}
