+++ airbyte-cdk/bulk/toolkits/extract-cdc/src/main/kotlin/io/airbyte/cdk/read/cdc/CdcContext.kt
@@ -2,7 +2,7 @@
  * Copyright (c) 2024 Airbyte, Inc., all rights reserved.
  */
 
-package io.airbyte.cdk.read
+package io.airbyte.cdk.read.cdc
 
 import io.airbyte.cdk.output.OutputConsumer
 import jakarta.inject.Singleton
