+++ airbyte-cdk/bulk/toolkits/extract-cdc/src/main/kotlin/io/airbyte/cdk/read/cdc/CdcGlobalLockResource.kt
@@ -2,9 +2,10 @@
  * Copyright (c) 2024 Airbyte, Inc., all rights reserved.
  */
 
-package io.airbyte.cdk.read
+package io.airbyte.cdk.read.cdc
 
 import io.airbyte.cdk.command.SourceConfiguration
+import io.airbyte.cdk.read.GlobalLockResource
 import io.micronaut.context.annotation.Replaces
 import jakarta.inject.Singleton
 import java.util.concurrent.atomic.AtomicBoolean
