+++ airbyte-cdk/bulk/toolkits/extract-cdc/src/main/kotlin/io/airbyte/cdk/read/cdc/CdcPartitionReader.kt
@@ -2,13 +2,11 @@
  * Copyright (c) 2024 Airbyte, Inc., all rights reserved.
  */
 
-package io.airbyte.cdk.cdc
+package io.airbyte.cdk.read.cdc
 
 import com.fasterxml.jackson.databind.ObjectMapper
 import io.airbyte.cdk.command.OpaqueStateValue
-import io.airbyte.cdk.read.CdcContext
 import io.airbyte.cdk.read.ConcurrencyResource
-import io.airbyte.cdk.read.DebeziumRecord
 import io.airbyte.cdk.read.PartitionReadCheckpoint
 import io.airbyte.cdk.read.PartitionReader
 import io.airbyte.cdk.util.Jsons
