+++ airbyte-cdk/bulk/toolkits/extract-cdc/src/main/kotlin/io/airbyte/cdk/read/cdc/InitialCdcStateCreator.kt
@@ -2,7 +2,7 @@
  * Copyright (c) 2024 Airbyte, Inc., all rights reserved.
  */
 
-package io.airbyte.cdk.read
+package io.airbyte.cdk.read.cdc
 
 import io.airbyte.cdk.command.OpaqueStateValue
 import io.debezium.engine.ChangeEvent
